[CR-2] clean code bloc and add stream location
