+++ lib/data/api/user_location_api.dart
@@ -1,0 +1,13 @@
+import 'dart:async';
+import 'package:location/location.dart';
+
+
+class UserLocationAPI {
+  Location location = Location();
+
+  UserLocationAPI();
+
+  Stream<LocationData> getLocation() {
+    return location.onLocationChanged;
+  }
+}
