+++ lib/domain/blocs/home_bloc/home_event.dart
@@ -1,4 +1,5 @@
 import 'package:equatable/equatable.dart';
+import 'package:location/location.dart';
 
 abstract class HomeEvent extends Equatable {
   const HomeEvent();
@@ -7,11 +8,26 @@ abstract class HomeEvent extends Equatable {
   List<Object> get props => [];
 }
 
-class HomeNewDataUserSpeedAndDistanceEvent extends HomeEvent {
-  final int speed;
-  final double totalDistance;
+class HomeStartedCurrentLocationEvent extends HomeEvent {
 
-  const HomeNewDataUserSpeedAndDistanceEvent({required this.speed, required this.totalDistance});
+  const HomeStartedCurrentLocationEvent();
+
+  @override
+  List<Object> get props => [];
+}
+
+class HomeNewDataUserSpeedAndDistanceEvent extends HomeEvent {
+  final LocationData currentLocation;
+
+  const HomeNewDataUserSpeedAndDistanceEvent({required this.currentLocation});
+
+  @override
+  List<Object> get props => [currentLocation];
+}
+
+class HomeResumedCurrentLocationEvent extends HomeEvent {
+
+  const HomeResumedCurrentLocationEvent();
 
   @override
   List<Object> get props => [];
