+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:car_helper_gps/presentation/page/home_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'data/api/user_location_api.dart';
 import 'domain/blocs/home_bloc/home_bloc.dart';
 
 
@@ -16,7 +17,7 @@ class MyApp extends StatelessWidget {
     return MultiBlocProvider(
       providers: [
         BlocProvider<HomeBloc>(
-          create: (context) => HomeBloc(),
+          create: (context) => HomeBloc(userLocation: UserLocationAPI()),
         ),
       ],
       child: MaterialApp(
